refactor: 폴더 정리, 타입과 상수 분리
