+++ server-toss/src/modules/drawing/service/save-drawing.service.ts
@@ -7,7 +7,7 @@ import { DrawingRepository } from "../drawing.repository";
 import { InjectRepository } from "@mikro-orm/nestjs";
 import { Transactional } from "@mikro-orm/decorators/legacy";
 import { PointService } from "src/modules/point/point.service";
-import { PointReason } from "src/modules/point/point-log.entity";
+import { PointReason } from "src/modules/point/entitiy/point-log.entity";
 
 @Injectable()
 export class SaveDrawingService {
