+++ server-toss/src/modules/drawing/test/save-drawing.service.test.ts
@@ -23,7 +23,7 @@ import { RankingRepository } from "../../ranking/ranking.repository";
 import { RankingService } from "../../ranking/ranking.service";
 import { SaveDrawingService } from "../service/save-drawing.service";
 import { PointService } from "src/modules/point/point.service";
-import { PointReason } from "src/modules/point/point-log.entity";
+import { PointReason } from "src/modules/point/entitiy/point-log.entity";
 
 describe("SaveDrawingService", () => {
   let orm: MikroORM;
