+++ server-toss/src/modules/point/point.contants.ts
@@ -1,0 +1,7 @@
+export const PROMOTION_AMOUNT = 2;
+export const PROMOTION_MAX_RETRIES = 3;
+export const PURGE_BATCH_SIZE = 100;
+export const SUCCEEDED_RETENTION_DAYS = 7;
+export const FAILED_RETENTION_DAYS = 30;
+export const DAILY_DRAWING_PROMOTION_LIMIT = 2;
+export const DAILY_LIMIT_EXCEEDED_MESSAGE = "일일 지급 한도 초과";
