+++ server-toss/src/modules/point/point.module.ts
@@ -1,10 +1,10 @@
 import { MikroOrmModule } from "@mikro-orm/nestjs";
 import { Module } from "@nestjs/common";
-import { PointGrantRequest } from "./point-grant-request.entity";
-import { PointLog } from "./point-log.entity";
+import { PointGrantRequest } from "./entitiy/point-grant-request.entity";
+import { PointLog } from "./entitiy/point-log.entity";
 import { PointService } from "./point.service";
-import { PointGrantScheduler } from "./point-grant.scheduler";
-import { PointGrantPurgeScheduler } from "./point-grant-purge.scheduler";
+import { PointGrantScheduler } from "./scheduler/point-grant.scheduler";
+import { PointGrantPurgeScheduler } from "./scheduler/point-grant-purge.scheduler";
 
 @Module({
   imports: [MikroOrmModule.forFeature([PointLog, PointGrantRequest])],
