+++ server-toss/src/modules/point/point.types.ts
@@ -1,0 +1,4 @@
+export type GrantEligibilityDecision =
+  | { decision: "PROCEED" }
+  | { decision: "FAIL"; reason: string }
+  | { decision: "RETRY" };
