+++ server-toss/src/modules/point/scheduler/point-grant.scheduler.ts
@@ -1,5 +1,5 @@
 import { Injectable } from "@nestjs/common";
-import { PointService } from "./point.service";
+import { PointService } from "../point.service";
 import { Cron, CronExpression } from "@nestjs/schedule";
 
 @Injectable()
