+++ server-toss/src/seeders/DatabaseSeeder.ts
@@ -3,7 +3,10 @@ import { Seeder } from "@mikro-orm/seeder";
 import { getTodayKst } from "src/common/util/today";
 import { AdType, AdView } from "src/modules/chance/ad-view.entity";
 import { Drawing } from "src/modules/drawing/drawing.entity";
-import { PointLog, PointReason } from "src/modules/point/point-log.entity";
+import {
+  PointLog,
+  PointReason,
+} from "src/modules/point/entitiy/point-log.entity";
 import { DailyPrompt } from "src/modules/prompt/daily-prompt.entity";
 import { Prompt } from "src/modules/prompt/prompt.entity";
 import { Ranking } from "src/modules/ranking/ranking.entity";
